Tests default to MySQL strict mode, binding based on php type.
This change breaks how binding works. Previously everything was bound to
single quote encapsulated tokens. This fails in strict mode. Now binding
relies on the php type of a variable for how it's bound into the query.
This means that a string will be bound to a string and escaped, an int
or float will be added without quoting and a bool will be represented
with a 1/0.