Force different IDs to expose cache bug
The test that now fails (`IncludesTest::testIncludesHitsCache()`) was
only passing because there was a 1-to-1 mapping between the IDs of
related entities. I.e. Hero with ID=1 corresponds to SecretIdentity with
ID=1, etc.

This forces SecretIdentity IDs to start at 100, which exposes the bug
discussed in https://github.com/lox/pheasant/issues/109.